==> scripts/00_load_packages.R <==
# load all libraries needed in the project

library(tidyverse)
library(skimr)
library(janitor)
library(caret)
library(ranger)
library(randomForest)
library(vip)        # model interpretation / variable importance plots
library(e1071)
library(dataedu)
